\name{RPathVisio-package}
\alias{RPathVisio-package}
\alias{RPathVisio}
\docType{package}
\title{
R package for PathVisio
}
\description{ 
Directly call functions of PathVisio from R using the PathVisioRPC server, to create and edit biological pathways, visualize data on pathways, perform pathway statistics and get a list of participating genes/metabolites/reactions etc. in a pathway.
}
\details{
This package wraps the PathVisioRPC calls with a R specific API.
}
\author{
Anwesha Bohler <anwesha.dutta@maastrichtuniversity.nl> \\
Egon Willighagen <egon.willighagen@gmail.com> \\
Christ Leemans <C.Leemans@student.maastrichtuniversity.nl>
}
\keyword{ package }
\keyword{ PathVisio }
\keyword{ XML-RPC }

